Update Repository and Data Sources

+++ app/src/main/java/com/bangkit/laporaja/data/LaporAjaRepositoryInterface.kt
@@ -5,6 +5,6 @@ import kotlinx.coroutines.flow.Flow
 
 interface LaporAjaRepositoryInterface {
     fun getRecentReports(): Flow<List<Report>>
-    fun getUserReports(userId: Int) : Flow<List<Report>>
-    fun getReportDetails(userId: Int, reportId: Int) : Flow<Report>
+    fun getUserReports(userId: Long) : Flow<List<Report>>
+    fun getReportDetails(userId: Long, reportId: Long) : Flow<Report>
 }
